Renamed 'store.ts'
