+++ src/components/Header/index.tsx
@@ -1,6 +1,6 @@
 import Button from '@/components/Button';
 import { Link } from 'react-router-dom';
-import { useStoreDispatch } from '@/store/store';
+import { useStoreDispatch } from '@/store';
 import { toClearState } from '@/store/cards';
 import styles from '@/components/Header/header.module.scss';
 import { ROUTES } from '@/store/routes';
