+++ src/components/modal/Input/index.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import styles from '@/components/modal/Input/input.module.scss';
 import { InputProps } from '@/store/cards/cards.types';
 import { useSelector } from 'react-redux';
-import { RootState } from '@/store/store';
+import { RootState } from '@/store';
 import classNames from 'classnames';
 
 const Input: FC<InputProps> = (props) => {
