+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from '@/App.tsx';
 import '@/index.scss';
-import { store } from '@/store/store.ts';
+import { store } from '@/store';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
